backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic-settings>=2.6.0
sqlalchemy>=2.0.36
psycopg2-binary>=2.9.10
redis>=5.2.0
aio-pika>=9.5.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-multipart>=0.0.12
pytest>=8.3.3
httpx>=0.27.2
email-validator>=2.0.0
python-dateutil
